add existsById in Delete and Update functions

+++ Program.cs
@@ -56,16 +56,32 @@ if(modelName == "Computer")
         var ram = args[3];
         var processor = args[4];
 
-        var computer = new Computer(id, ram, processor);
-        computerRepository.Update(computer);
-        Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
+        if(computerRepository.existsById(id))
+        {
+            var computer = new Computer(id, ram, processor);
+            computerRepository.Update(computer);
+            Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
+        }
+
+        else
+        {
+            Console.WriteLine($"O computador {id} não existe");
+        }
     }
     
     if(modelAction == "Delete")
     {
         var id = Convert.ToInt32(args[2]);
 
-        computerRepository.Delete(id);
+        if(computerRepository.existsById(id))
+        {
+            computerRepository.Delete(id);
+        }
+
+        else
+        {
+            Console.WriteLine($"O computador {id} não existe");
+        }
     }
 }
